Add files via upload
Description:

The "ATM Management System" is a Java-based console application that replicates the functionalities of an ATM commonly found in cities. It allows users to perform various financial transactions after authentication.

Key Features:

1. **Authentication**: Users must enter their user ID and PIN to access the ATM functionalities, ensuring security.

2. **Transactions History**: Users can view their recent transaction history, providing them with a record of their financial activities.

3. **Withdraw**: Users can withdraw cash from their accounts by specifying the desired amount, with the system verifying fund availability.

4. **Deposit**: Users can deposit cash into their accounts, with the deposited amount being added to their balance.

5. **Transfer**: Users can transfer funds between their accounts or to another account within the bank securely.

6. **Quit**: Users can exit the ATM system when they're done with their transactions.

The project consists of five classes to manage different aspects of ATM functionality, offering a straightforward solution for handling financial transactions securely. It provides users with a convenient and reliable platform for managing their finances effectively.